api/requirements.txt
asn1crypto==0.24.0
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.1.2
datapunt-authorization-django==0.2.18
datapunt-authorization-levels==0.1.3
Django==2.0.1
django-braces==1.12.0
django-extensions==1.9.9
django-filter==1.1.0
django-rest-swagger==2.1.2
djangorestframework==3.7.7
djangorestframework-gis==0.12
factory-boy==2.10.0
drf-extensions==0.3.1
git+https://github.com/Amsterdam/drf_amsterdam.git#egg=drf_amsterdam
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
openapi-codec==1.3.2
psycopg2-binary==2.7.4
pycparser==2.18
PyJWT==1.5.3
pytz==2017.3
requests==2.18.4
simplejson==3.13.2
six==1.11.0
typing==3.6.2
uritemplate==3.0.0
urllib3==1.22
